@@ -1,6 +1,6 @@
 {
   "name": "custom-functions-metadata-plugin",
-  "version": "1.0.32",
+  "version": "1.2.9",
   "author": "Office Dev",
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
@@ -18,23 +18,25 @@
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts"
   },
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "test": "echo \"No tests.\""
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
+    "test": "echo \"No tests.\"",
+    "watch": "tsc -p tsconfig.json -w"
   },
   "dependencies": {
-    "custom-functions-metadata": "^1.0.32",
-    "es6-promise": "^4.2.8",
-    "webpack": "^4.42.1"
+    "custom-functions-metadata": "^1.2.9",
+    "loader-utils": "^2.0.0",
+    "webpack": "^5.24.4"
   },
   "devDependencies": {
-    "@types/node": "^12.12.34",
-    "@types/webpack": "^4.41.10",
-    "concurrently": "^5.2.0",
-    "rimraf": "^2.7.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3"
+    "@types/loader-utils": "^2.0.2",
+    "@types/node": "^14.17.2",
+    "concurrently": "^6.2.2",
+    "office-addin-lint": "^2.0.3",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.3.5"
   },
   "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
