@@ -1,19 +1,18 @@
 {
   "name": "custom-functions-metadata",
-  "version": "1.0.32",
+  "version": "1.2.9",
   "description": "Generate metadata for Excel Custom Functions.",
   "main": "./lib/main.js",
   "scripts": {
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
     "cli": "node lib/cli.js",
     "lint": "office-addin-lint check",
     "lint:fix": "office-addin-lint fix",
+    "package": "node ./lib/cli.js generate",
     "prettier": "office-addin-lint prettier",
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "start": "rimraf lib && concurrently \"npm run tsc-watch\"",
+    "start": "rimraf lib && concurrently \"npm run watch\"",
     "test": "mocha -r ts-node/register test/src/*.ts",
-    "package": "node ./lib/cli.js generate"
+    "watch": "tsc -p tsconfig.json -w"
   },
   "author": "Office Dev",
   "license": "MIT",
@@ -27,26 +26,24 @@
   ],
   "dependencies": {
     "assert": "^1.5.0",
-    "commander": "^2.20.3",
-    "nconf": "^0.10.0",
-    "office-addin-cli": "^1.0.12",
+    "commander": "^6.2.0",
+    "nconf": "^0.12.0",
+    "office-addin-usage-data": "^1.4.9",
     "xregexp": "^4.3.0"
   },
   "devDependencies": {
     "@types/assert": "^1.4.6",
-    "@types/commander": "^2.12.2",
-    "@types/es6-promise": "^3.3.0",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^12.12.34",
-    "@types/node-fetch": "^2.5.5",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
     "@types/xregexp": "^3.0.30",
-    "concurrently": "^3.6.1",
-    "mocha": "^7.1.1",
-    "office-addin-lint": "^1.0.25",
-    "office-addin-prettier-config": "^1.0.13",
-    "rimraf": "^2.7.1",
-    "ts-node": "^8.8.1",
-    "typescript": "^3.8.3"
+    "concurrently": "^6.2.2",
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^2.0.3",
+    "office-addin-prettier-config": "^1.1.7",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
