@@ -1,21 +1,31 @@
 {
   "name": "eslint-config-office-addins",
-  "version": "1.0.17",
+  "version": "1.4.9",
   "description": "Provides an eslint configuration for Office Add-ins.",
   "main": "./lib/main.js",
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
     "test": "echo \"No tests\""
   },
   "keywords": [
     "eslint"
   ],
+  "dependencies": {
+    "eslint": "^7.32.0",
+    "eslint-plugin-office-addins": "^2.0.3"
+  },
+  "peerDependencies": {
+    "prettier": "^2.1.2"
+  },
   "devDependencies": {
-    "@types/node": "^12.12.34",
-    "concurrently": "^3.6.1",
+    "@types/node": "^14.17.2",
+    "concurrently": "^6.2.2",
+    "office-addin-lint": "^2.0.3",
     "rimraf": "^2.7.1",
-    "typescript": "^3.8.3"
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
