@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-office-addins",
-  "version": "0.1.8",
+  "version": "2.0.3",
   "description": "eslint plugin for office-addins",
   "keywords": [
     "eslint",
@@ -10,21 +10,32 @@
   "author": "OfficeDev",
   "main": "lib/main.js",
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "test": "echo \"No tests\""
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "lint": "eslint src/**/*.{ts,tsx,js,jsx} --rulesdir lib/rules/",
+    "lint:fix": "eslint src/**/*.{ts,tsx,js,jsx} --rulesdir lib/rules/ --fix",
+    "test": "jest"
   },
   "dependencies": {
-    "eslint": "^6.8.0",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-native": "^3.8.1",
-    "requireindex": "~1.1.0"
+    "@typescript-eslint/eslint-plugin": "^4.31.0",
+    "@typescript-eslint/experimental-utils": "^4.31.0",
+    "@typescript-eslint/parser": "^4.31.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.25.1",
+    "eslint-plugin-react-native": "^3.11.0",
+    "office-addin-prettier-config": "^1.1.7",
+    "prettier": "^2.4.0",
+    "requireindex": "~1.2.0",
+    "typescript": "^4.4.2"
   },
   "devDependencies": {
-    "@types/node": "^12.12.34",
-    "concurrently": "^3.6.1",
-    "rimraf": "^2.6.2",
-    "typescript": "^3.8.3"
+    "@types/jest": "^27.0.1",
+    "@types/node": "^14.17.2",
+    "concurrently": "^6.2.1",
+    "jest": "^27.2.4",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^27.0.5"
   },
   "engines": {
     "node": ">=0.10.0"
