@@ -1,16 +1,18 @@
 {
   "name": "office-addin-debugging",
-  "version": "3.0.33",
+  "version": "4.3.14",
   "description": "For debugging Office Add-ins.",
   "main": "./lib/main.js",
   "scripts": {
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
     "cli": "node lib/cli.js",
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "start": "rimraf lib && concurrently \"npm run tsc-watch\"",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "node-debugger": "office-addin-node-debugger",
+    "prettier": "office-addin-lint prettier",
+    "start": "rimraf lib && concurrently \"npm run watch\"",
     "test": "mocha -r ts-node/register test/**/*.ts",
-    "node-debugger": "office-addin-node-debugger"
+    "watch": "tsc -p tsconfig.json -w"
   },
   "author": "Office Dev",
   "license": "MIT",
@@ -22,30 +24,29 @@
   ],
   "dependencies": {
     "child_process": "^1.0.2",
-    "commander": "^2.20.3",
-    "node-fetch": "^2.2.0",
-    "office-addin-cli": "^1.0.12",
-    "office-addin-dev-certs": "^1.5.4",
-    "office-addin-dev-settings": "^1.8.5",
-    "office-addin-manifest": "^1.5.6",
-    "office-addin-node-debugger": "^0.4.31",
-    "office-addin-usage-data": "^1.1.6"
+    "commander": "^6.2.0",
+    "node-fetch": "^2.6.1",
+    "office-addin-cli": "^1.3.8",
+    "office-addin-dev-certs": "^1.7.10",
+    "office-addin-dev-settings": "^1.11.4",
+    "office-addin-manifest": "^1.8.1",
+    "office-addin-node-debugger": "^0.6.9",
+    "office-addin-usage-data": "^1.4.9"
   },
   "devDependencies": {
-    "@types/es6-promise": "0.0.32",
-    "@types/express": "^4.17.3",
+    "@types/express": "^4.17.13",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^12.12.34",
-    "@types/node-fetch": "^2.5.5",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
     "@types/ws": "^6.0.4",
-    "concurrently": "^3.5.0",
-    "express": "^4.16.4",
-    "mocha": "^7.1.1",
+    "concurrently": "^6.2.2",
+    "express": "^4.17.3",
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^2.0.3",
     "rimraf": "^2.7.1",
-    "ts-node": "^8.8.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3",
-    "ws": "^6.1.0"
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
+    "ws": "^7.4.6"
   },
   "repository": {
     "type": "git",
@@ -54,5 +55,6 @@
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
   },
+  "prettier": "office-addin-prettier-config",
   "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
