@@ -1,11 +1,14 @@
 {
   "name": "office-addin-dev-settings",
-  "version": "1.8.5",
+  "version": "1.11.4",
   "description": "Configure developer settings for Office Add-ins.",
   "main": "./lib/main.js",
   "scripts": {
-    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "build": "rimraf lib && copyfiles \"src/DevXTool.exe\" -u 1 lib && concurrently \"tsc -p tsconfig.json\"",
     "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
     "test": "mocha -r ts-node/register test/**/*.ts",
     "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\""
   },
@@ -19,36 +22,37 @@
     "Office Add-in"
   ],
   "dependencies": {
+    "@microsoft/teamsfx-cli": "0.12.1",
     "child_process": "^1.0.2",
-    "commander": "^2.20.3",
-    "fs-extra": "^7.0.1",
-    "inquirer": "^7.1.0",
-    "jszip": "^3.2.2",
+    "commander": "^6.2.0",
+    "fs-extra": "^9.0.1",
+    "inquirer": "^7.3.3",
+    "jszip": "^3.7.1",
     "junk": "^3.1.0",
-    "office-addin-cli": "^1.0.12",
-    "office-addin-manifest": "^1.5.6",
+    "office-addin-manifest": "^1.8.1",
+    "office-addin-usage-data": "^1.4.9",
     "open": "^6.4.0",
     "whatwg-url": "^7.1.0",
     "winreg": "^1.2.4"
   },
   "devDependencies": {
-    "@types/es6-promise": "^3.3.0",
-    "@types/fs-extra": "^5.1.0",
+    "@types/fs-extra": "^9.0.1",
     "@types/inquirer": "^6.5.0",
-    "@types/jszip": "^3.1.7",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.12.34",
-    "@types/node-fetch": "^2.5.5",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/semver": "^7.3.4",
     "@types/sinon": "^7.5.2",
     "@types/whatwg-url": "^6.4.0",
     "@types/winreg": "^1.2.30",
-    "concurrently": "^3.6.1",
-    "mocha": "^7.1.1",
-    "rimraf": "^2.7.1",
+    "concurrently": "^6.2.2",
+    "copyfiles": "2.4.1",
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^2.0.3",
+    "rimraf": "^3.0.2",
     "sinon": "^7.5.0",
-    "ts-node": "^8.8.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3"
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
@@ -57,5 +61,6 @@
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
   },
+  "prettier": "office-addin-prettier-config",
   "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
