@@ -1,12 +1,14 @@
 {
   "name": "office-addin-lint",
-  "version": "1.0.25",
+  "version": "2.0.3",
   "description": "Provide eslint and prettier integration",
   "main": "./lib/main.js",
   "scripts": {
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
     "cli": "node lib/cli.js",
-    "tsc": "tsc -p tsconfig.json",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
+    "lint": "node lib/cli.js check",
+    "lint:fix": "node lib/cli.js fix",
+    "prettier": "node lib/cli.js prettier",
     "test": "mocha -r ts-node/register test/src/*.ts"
   },
   "author": "Office Dev",
@@ -22,30 +24,28 @@
     "Office Add-in"
   ],
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^2.26.0",
-    "@typescript-eslint/parser": "^2.26.0",
-    "commander": "^2.20.3",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
-    "eslint-plugin-prettier": "^3.1.2",
-    "office-addin-usage-data": "^1.1.6",
-    "prettier": "^1.19.1"
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.15.1",
+    "commander": "^6.2.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-office-addins": "^2.0.3",
+    "eslint-plugin-prettier": "^3.3.1",
+    "office-addin-prettier-config": "^1.1.7",
+    "office-addin-usage-data": "^1.4.9",
+    "prettier": "^2.2.1"
   },
   "devDependencies": {
     "@types/assert": "^1.4.6",
-    "@types/commander": "^2.12.2",
-    "@types/es6-promise": "^3.3.0",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^12.12.34",
-    "@types/node-fetch": "^2.5.5",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
     "assert": "^1.4.1",
-    "concurrently": "^3.6.1",
-    "mocha": "^7.1.1",
-    "office-addin-cli": "^1.0.12",
-    "rimraf": "^2.7.1",
-    "ts-node": "^8.8.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3"
+    "concurrently": "^6.2.2",
+    "mocha": "^9.1.1",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
@@ -57,28 +57,10 @@
   "prettier": "office-addin-prettier-config",
   "eslintConfig": {
     "plugins": [
-      "prettier",
-      "@typescript-eslint"
+      "office-addins"
     ],
-    "rules": {
-      "prettier/prettier": "error",
-      "no-eval": "error",
-      "no-delete-var": "warn",
-      "no-octal": "warn",
-      "no-inner-declarations": "warn"
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-      "ecmaVersion": 6,
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true
-      },
-      "project": "./tsconfig.json"
-    },
     "extends": [
-      "prettier",
-      "prettier/@typescript-eslint"
+      "plugin:office-addins/recommended"
     ]
   },
   "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
