@@ -1,11 +1,14 @@
 {
   "name": "office-addin-manifest",
-  "version": "1.5.6",
+  "version": "1.8.1",
   "description": "Read and modify Office Add-in manifest files.",
   "main": "./lib/main.js",
   "scripts": {
     "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
     "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
     "test": "mocha -r ts-node/register test/**/*.ts",
     "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\""
   },
@@ -19,32 +22,34 @@
     "Office Add-in"
   ],
   "dependencies": {
+    "@microsoft/teams-manifest": "^0.0.3",
     "chalk": "^2.4.2",
-    "commander": "^2.20.3",
-    "node-fetch": "^2.6.0",
-    "office-addin-cli": "^1.0.12",
-    "office-addin-usage-data": "^1.1.6",
+    "commander": "^6.2.0",
+    "fs-extra": "^7.0.1",
+    "jszip": "^3.7.1",
+    "node-fetch": "^2.6.1",
+    "office-addin-usage-data": "^1.4.9",
     "path": "^0.12.7",
-    "uuid": "^3.4.0",
+    "tslib": "^2.4.0",
+    "uuid": "^8.3.2",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@types/es6-promise": "0.0.32",
+    "@types/fs-extra": "^9.0.13",
     "@types/mocha": "^5.2.7",
-    "@types/node": "^12.12.34",
-    "@types/node-fetch": "^2.5.5",
-    "@types/uuid": "^3.4.8",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/uuid": "^8.3.0",
     "@types/validator": "^9.4.4",
     "@types/xml2js": "^0.4.5",
-    "concurrently": "^3.6.1",
+    "concurrently": "^6.2.2",
     "copy-dir": "^0.4.0",
-    "fs-extra": "^7.0.0",
-    "mocha": "^7.1.1",
-    "rimraf": "^2.7.1",
-    "ts-node": "^8.8.1",
-    "tslint": "^6.1.0",
-    "typescript": "^3.8.3",
-    "validator": "^10.11.0"
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^2.0.3",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
+    "validator": "^13.7.0"
   },
   "repository": {
     "type": "git",
@@ -53,5 +58,6 @@
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
   },
+  "prettier": "office-addin-prettier-config",
   "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
