@@ -6,16 +6,19 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-import * as child from 'child_process';
-import * as commander from 'commander';
-import { fork } from 'child_process';
-import * as path from "path";
-import WebSocket = require('ws');
+import * as child from "child_process";
+import { fork } from "child_process";
+import WebSocket = require("ws");
 
-export function run(host: string = "localhost", port: string = "8081", 
-  role: string = "debugger", debuggerName: string = "OfficeAddinDebugger") {
-    
-  const debuggerWorkerRelativePath: string = '\\debuggerWorker.js';
+/* global console, __dirname, setTimeout */
+
+export function run(
+  host: string = "localhost",
+  port: string = "8081",
+  role: string = "debugger",
+  debuggerName: string = "OfficeAddinDebugger"
+) {
+  const debuggerWorkerRelativePath: string = "\\debuggerWorker.js";
   const debuggerWorkerFullPath: string = `${__dirname}${debuggerWorkerRelativePath}`;
   const websocketRetryTimeout: number = 500;
 
@@ -24,13 +27,12 @@ export function run(host: string = "localhost", port: string = "8081",
     var worker: child.ChildProcess;
 
     function createJSRuntime(): void {
-      
       // This worker will run the application javascript code.
       worker = fork(`${debuggerWorkerFullPath}`, [], {
-        stdio: ['pipe', 'pipe', 'pipe', 'ipc'],
-        execArgv: ['--inspect']
+        stdio: ["pipe", "pipe", "pipe", "ipc"],
+        execArgv: ["--inspect"],
       });
-      worker.on('message', message => {
+      worker.on("message", (message) => {
         ws.send(JSON.stringify(message));
       });
     }
@@ -43,16 +45,16 @@ export function run(host: string = "localhost", port: string = "8081",
     }
 
     ws.onopen = () => {
-      console.log('Web socket opened...');
+      console.log("Web socket opened...");
     };
-    ws.onmessage = message => {
+    ws.onmessage = (message) => {
       if (!message.data) {
         return;
       }
 
       var object = JSON.parse(message.data.toString());
 
-      if (object.$event === 'client-disconnected') {
+      if (object.$event === "client-disconnected") {
         shutdownJSRuntime();
         return;
       }
@@ -60,26 +62,26 @@ export function run(host: string = "localhost", port: string = "8081",
         return;
       }
       // Special message that asks for a new JS runtime
-      if (object.method === 'prepareJSRuntime') {
+      if (object.method === "prepareJSRuntime") {
         shutdownJSRuntime();
         //console.clear();
 
         createJSRuntime();
         ws.send(JSON.stringify({ replyID: object.id }));
-      } else if (object.method === '$disconnected') {
+      } else if (object.method === "$disconnected") {
         shutdownJSRuntime();
       } else {
         worker.send(object);
       }
     };
-    ws.onclose = e => {
+    ws.onclose = (e) => {
       shutdownJSRuntime();
       if (e.reason) {
         console.log(`Web socket closed because the following reason: ${e.reason}`);
       }
       setTimeout(connectToDebuggerProxy, websocketRetryTimeout);
     };
-    ws.onerror = event => {
+    ws.onerror = (event) => {
       console.log(`${event.error}`);
     };
   }
