@@ -1,22 +1,31 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import * as commander from "commander";
-import { logErrorMessage } from "office-addin-cli";
-import * as generateMetadata from "./generate";
+import { writeFileSync } from "fs";
+import { logErrorMessage } from "office-addin-usage-data";
+import { generateCustomFunctionsMetadata } from "./generate";
 
-export async function generate(inputFile: string, outputFile: string) {
+/* global console */
+
+export async function generate(inputPath: string, outputPath: string) {
   try {
-    if (!inputFile) {
+    if (!inputPath) {
       throw new Error("You need to provide the path to the source file for custom functions.");
     }
-    if (!outputFile) {
-      throw new Error("You need to provide the path to the output file for the custom functions metadata.");
-    }
-    const results = await generateMetadata.generate(inputFile, outputFile);
+    const results = await generateCustomFunctionsMetadata(inputPath);
     if (results.errors.length > 0) {
       console.error("Errors found:");
-      results.errors.forEach(err => console.log(err));
+      results.errors.forEach((err) => console.log(err));
+    } else {
+      if (outputPath) {
+        try {
+          writeFileSync(outputPath, results.metadataJson);
+        } catch (err) {
+          throw new Error(`Cannot write to file: ${outputPath}.`);
+        }
+      } else {
+        console.log(results.metadataJson);
+      }
     }
   } catch (err) {
     logErrorMessage(err);
