@@ -3,7 +3,9 @@
 
 import * as fs from "fs";
 import * as ts from "typescript";
-import * as XRegExp from "xregexp";
+import XRegExp = require("xregexp");
+
+/* global console */
 
 export interface ICustomFunctionsMetadata {
   functions: IFunction[];
@@ -12,35 +14,37 @@ export interface ICustomFunctionsMetadata {
 export interface IFunction {
   id: string;
   name: string;
-  description: string;
-  helpUrl: string;
+  description?: string;
+  helpUrl?: string;
   parameters: IFunctionParameter[];
-  result: IFunctionResult;
-  options: IFunctionOptions;
+  result?: IFunctionResult;
+  options?: IFunctionOptions;
 }
 
 export interface IFunctionOptions {
-  cancelable: boolean;
-  requiresAddress: boolean;
-  stream: boolean;
-  volatile: boolean;
+  cancelable?: boolean;
+  requiresAddress?: boolean;
+  stream?: boolean;
+  volatile?: boolean;
+  requiresParameterAddresses?: boolean;
 }
 
 export interface IFunctionParameter {
   name: string;
   description?: string;
   type: string;
-  dimensionality: string;
-  optional: boolean;
-  repeating: boolean;
+  dimensionality?: string;
+  optional?: boolean;
+  repeating?: boolean;
 }
 
 export interface IFunctionResult {
-  type: string;
-  dimensionality: string;
+  type?: string;
+  dimensionality?: string;
 }
 
 export interface IGenerateResult {
+  metadataJson: string;
   associate: IAssociate[];
   errors: string[];
 }
@@ -80,23 +84,29 @@ interface IGetParametersArguments {
   extra: IFunctionExtras;
   jsDocParamInfo: { [key: string]: string };
   jsDocParamOptionalInfo: { [key: string]: string };
-  jsDocParamTypeInfo: { [key: string]: string };
+  jsDocParamTypeInfo: { [key: string]: IJsDocParamType };
   parametersToParse: ts.ParameterDeclaration[];
+}
+
+interface IJsDocParamType {
+  type: string;
+  returnType: string;
+  dimensionality: string;
 }
 
 const CUSTOM_FUNCTION = "customfunction"; // case insensitive @CustomFunction tag to identify custom functions in JSDoc
 const HELPURL_PARAM = "helpurl";
 const VOLATILE = "volatile";
 const STREAMING = "streaming";
-const RETURN = "return";
 const CANCELABLE = "cancelable";
 const REQUIRESADDRESS = "requiresaddress";
+const REQUIRESPARAMETERADDRESSES = "requiresparameteraddresses";
 
 const TYPE_MAPPINGS_SIMPLE = {
   [ts.SyntaxKind.NumberKeyword]: "number",
   [ts.SyntaxKind.StringKeyword]: "string",
   [ts.SyntaxKind.BooleanKeyword]: "boolean",
-  [ts.SyntaxKind.AnyKeyword]: "any"
+  [ts.SyntaxKind.AnyKeyword]: "any",
 };
 
 const TYPE_MAPPINGS = {
@@ -108,39 +118,17 @@ const TYPE_MAPPINGS = {
   [ts.SyntaxKind.TupleType]: "any",
   [ts.SyntaxKind.EnumKeyword]: "any",
   [ts.SyntaxKind.ObjectKeyword]: "any",
-  [ts.SyntaxKind.VoidKeyword]: "any"
-};
-
-const TYPE_MAPPINGS_COMMENT = {
-  ["number"]: "number",
-  ["string"]: "string",
-  ["boolean"]: "boolean",
-  ["any"]: "any",
-  ["number[]"]: "number",
-  ["number[][]"]: "number",
-  ["number[][][]"]: "number",
-  ["string[]"]: "string",
-  ["string[][]"]: "string",
-  ["string[][][]"]: "string",
-  ["boolean[]"]: "boolean",
-  ["boolean[][]"]: "boolean",
-  ["boolean[][][]"]: "boolean"
+  [ts.SyntaxKind.VoidKeyword]: "any",
 };
 
 const TYPE_CUSTOM_FUNCTIONS_STREAMING = {
-  ["customfunctions.streaminghandler<string>"]: "string",
-  ["customfunctions.streaminghandler<number>"]: "number",
-  ["customfunctions.streaminghandler<boolean>"]: "boolean",
-  ["customfunctions.streaminghandler<any>"]: "any",
-  ["customfunctions.streaminginvocation<string>"]: "string",
-  ["customfunctions.streaminginvocation<number>"]: "number",
-  ["customfunctions.streaminginvocation<boolean>"]: "boolean",
-  ["customfunctions.streaminginvocation<any>"]: "any"
+  ["customfunctions.streaminghandler"]: 1,
+  ["customfunctions.streaminginvocation"]: 2,
 };
 
 const TYPE_CUSTOM_FUNCTION_CANCELABLE = {
   ["customfunctions.cancelablehandler"]: 1,
-  ["customfunctions.cancelableinvocation"]: 2
+  ["customfunctions.cancelableinvocation"]: 2,
 };
 const TYPE_CUSTOM_FUNCTION_INVOCATION = "customfunctions.invocation";
 
@@ -151,47 +139,35 @@ type CustomFunctionsSchemaDimensionality = "invalid" | "scalar" | "matrix";
  * @param inputFile - File that contains the custom functions
  * @param outputFileName - Name of the file to create (i.e functions.json)
  */
-export async function generate(
+export async function generateCustomFunctionsMetadata(
   inputFile: string,
-  outputFileName: string,
   wantConsoleOutput: boolean = false
 ): Promise<IGenerateResult> {
-  const errors: string[] = [];
-  const associate: IAssociate[] = [];
   const generateResults: IGenerateResult = {
-    associate,
-    errors
+    metadataJson: "",
+    associate: [],
+    errors: [],
   };
 
   if (fs.existsSync(inputFile)) {
     const sourceCode = fs.readFileSync(inputFile, "utf-8");
     const parseTreeResult: IParseTreeResult = parseTree(sourceCode, inputFile);
-    parseTreeResult.extras.forEach(extra => extra.errors.forEach(err => errors.push(err)));
-    generateResults.associate = [...parseTreeResult.associate];
+    parseTreeResult.extras.forEach((extra) => extra.errors.forEach((err) => generateResults.errors.push(err)));
 
-    if (errors.length === 0) {
-      const json = JSON.stringify({ functions: parseTreeResult.functions }, null, 4);
-
-      try {
-        fs.writeFileSync(outputFileName, json);
-
-        if (wantConsoleOutput) {
-          console.log(`${outputFileName} created for file: ${inputFile}`);
-        }
-      } catch (err) {
-        if (wantConsoleOutput) {
-          console.error(err);
-        }
-        throw new Error(`Error writing: ${outputFileName} : ${err}`);
+    if (generateResults.errors.length > 0) {
+      if (wantConsoleOutput) {
+        console.error("Errors in file: " + inputFile);
+        generateResults.errors.forEach((err) => console.error(err));
       }
-    } else if (wantConsoleOutput) {
-      console.error("Errors in file: " + inputFile);
-      errors.forEach(err => console.error(err));
+    } else {
+      generateResults.metadataJson = JSON.stringify({ functions: parseTreeResult.functions }, null, 4);
+      generateResults.associate = [...parseTreeResult.associate];
     }
   } else {
     throw new Error(`File not found: ${inputFile}`);
   }
-  return Promise.resolve(generateResults);
+
+  return generateResults;
 }
 
 /**
@@ -200,7 +176,7 @@ export async function generate(
  * @param sourceFileName source code file name or path
  * @param parseTreeOptions options to enable or disable
  */
-export function parseTree(sourceCode: string, sourceFileName: string, parseTreeOptions?: IOptions): IParseTreeResult {
+export function parseTree(sourceCode: string, sourceFileName: string): IParseTreeResult {
   const associate: IAssociate[] = [];
   const functions: IFunction[] = [];
   const extras: IFunctionExtras[] = [];
@@ -216,7 +192,7 @@ export function parseTree(sourceCode: string, sourceFileName: string, parseTreeO
   const parseTreeResult: IParseTreeResult = {
     associate,
     extras,
-    functions
+    functions,
   };
   return parseTreeResult;
 
@@ -245,7 +221,7 @@ export function parseTree(sourceCode: string, sourceFileName: string, parseTreeO
         if (isCustomFunction(functionDeclaration)) {
           const extra: IFunctionExtras = {
             errors: functionErrors,
-            javascriptFunctionName: functionName
+            javascriptFunctionName: functionName,
           };
           const idName = getTagComment(functionDeclaration, CUSTOM_FUNCTION);
           const idNameArray = idName.split(" ");
@@ -269,7 +245,7 @@ export function parseTree(sourceCode: string, sourceFileName: string, parseTreeO
             jsDocParamInfo,
             jsDocParamOptionalInfo,
             jsDocParamTypeInfo,
-            parametersToParse
+            parametersToParse,
           };
           const parameters = getParameters(parameterItems);
 
@@ -322,10 +298,16 @@ export function parseTree(sourceCode: string, sourceFileName: string, parseTreeO
             name,
             options,
             parameters,
-            result
+            result,
           };
 
-          if (!options.cancelable && !options.requiresAddress && !options.stream && !options.volatile) {
+          if (
+            !options.cancelable &&
+            !options.requiresAddress &&
+            !options.stream &&
+            !options.volatile &&
+            !options.requiresParameterAddresses
+          ) {
             delete functionMetadata.options;
           } else {
             if (!options.cancelable) {
@@ -342,6 +324,10 @@ export function parseTree(sourceCode: string, sourceFileName: string, parseTreeO
 
             if (!options.volatile) {
               delete options.volatile;
+            }
+
+            if (!options.requiresParameterAddresses) {
+              delete options.requiresParameterAddresses;
             }
           }
 
@@ -482,20 +468,22 @@ function getOptions(
     cancelable: isCancelableTag(func, isCancelableFunction),
     requiresAddress: isAddressRequired(func),
     stream: isStreaming(func, isStreamingFunction),
-    volatile: isVolatile(func)
+    volatile: isVolatile(func),
+    requiresParameterAddresses: isRequiresParameterAddresses(func),
   };
 
-  if (optionsItem.requiresAddress) {
+  if (optionsItem.requiresAddress || optionsItem.requiresParameterAddresses) {
+    let errorParam: string = optionsItem.requiresAddress ? "@requiresAddress" : "@requiresParameterAddresses";
+
     if (!isStreamingFunction && !isCancelableFunction && !isInvocationFunction) {
       const functionPosition = getPosition(func, func.parameters.end);
-      const errorString =
-        "Since @requiresAddress is present, the last function parameter should be of type CustomFunctions.Invocation :";
+      const errorString = `Since ${errorParam} is present, the last function parameter should be of type CustomFunctions.Invocation :`;
       extra.errors.push(logError(errorString, functionPosition));
     }
 
     if (isStreamingFunction) {
       const functionPosition = getPosition(func);
-      const errorString = "@requiresAddress cannot be used with @streaming.";
+      const errorString = `${errorParam} cannot be used with @streaming.`;
       extra.errors.push(logError(errorString, functionPosition));
     }
   }
@@ -513,7 +501,7 @@ function getResults(
   func: ts.FunctionDeclaration,
   isStreamingFunction: boolean,
   lastParameter: ts.ParameterDeclaration,
-  jsDocParamTypeInfo: { [key: string]: string },
+  jsDocParamTypeInfo: { [key: string]: IJsDocParamType },
   extra: IFunctionExtras,
   enumList: string[]
 ): IFunctionResult {
@@ -521,7 +509,7 @@ function getResults(
   let resultDim = "scalar";
   const defaultResultItem: IFunctionResult = {
     dimensionality: resultDim,
-    type: resultType
+    type: resultType,
   };
 
   const lastParameterPosition = getPosition(lastParameter);
@@ -534,10 +522,11 @@ function getResults(
       const name = (lastParameter.name as ts.Identifier).text;
       const ptype = jsDocParamTypeInfo[name];
       // @ts-ignore
-      resultType = TYPE_CUSTOM_FUNCTIONS_STREAMING[ptype.toLocaleLowerCase()];
+      resultType = ptype.returnType;
+      resultDim = ptype.dimensionality;
       const paramResultItem: IFunctionResult = {
         dimensionality: resultDim,
-        type: resultType
+        type: resultType,
       };
 
       if (paramResultItem.dimensionality === "scalar") {
@@ -620,7 +609,7 @@ function getResults(
 
   const resultItem: IFunctionResult = {
     dimensionality: resultDim,
-    type: resultType
+    type: resultType,
   };
 
   // Only return dimensionality = matrix.  Default assumed scalar
@@ -643,7 +632,7 @@ function getResults(
  */
 function getParameters(parameterItem: IGetParametersArguments): IFunctionParameter[] {
   const parameterMetadata: IFunctionParameter[] = [];
-  const parameters = parameterItem.parametersToParse
+  parameterItem.parametersToParse
     .map((p: ts.ParameterDeclaration) => {
       const parameterPosition = getPosition(p);
       // Get type node of parameter from typescript
@@ -670,7 +659,7 @@ function getParameters(parameterItem: IGetParametersArguments): IFunctionParamet
         name,
         optional: getParamOptional(p, parameterItem.jsDocParamOptionalInfo),
         repeating: isRepeatingParameter(typeNode),
-        type: ptype
+        type: ptype,
       };
 
       // Only return dimensionality = matrix.  Default assumed scalar
@@ -695,7 +684,7 @@ function getParameters(parameterItem: IGetParametersArguments): IFunctionParamet
 
       parameterMetadata.push(pMetadataItem);
     })
-    .filter(meta => meta);
+    .filter((meta) => meta);
 
   return parameterMetadata;
 }
@@ -751,9 +740,9 @@ function findTag(node: ts.Node, tagName: string): ts.JSDocTag | undefined {
 /**
  * If a node contains the named tag, returns the tag comment, otherwise returns "".
  */
-function getTagComment(node: ts.Node, tagName: string) {
+function getTagComment(node: ts.Node, tagName: string): string {
   const tag = findTag(node, tagName);
-  return tag && tag.comment ? tag.comment : "";
+  return tag?.comment?.toString() || "";
 }
 
 /**
@@ -789,6 +778,14 @@ function isAddressRequired(node: ts.Node): boolean {
   return hasTag(node, REQUIRESADDRESS);
 }
 
+/**
+ * Returns true if RequiresParameterAddresses tag found in comments
+ * @param node jsDocs node
+ */
+function isRequiresParameterAddresses(node: ts.Node): boolean {
+  return hasTag(node, REQUIRESPARAMETERADDRESSES);
+}
+
 function containsTag(tag: ts.JSDocTag, tagName: string): boolean {
   return (tag.tagName.escapedText as string).toLowerCase() === tagName;
 }
@@ -812,28 +809,6 @@ function isCancelableTag(node: ts.Node, cancelableFunction: boolean): boolean {
 }
 
 /**
- * Returns return type of function from comments
- * @param node - jsDocs node
- */
-function getReturnType(node: ts.Node): string {
-  let type = "any";
-  ts.getJSDocTags(node).forEach((tag: ts.JSDocTag) => {
-    if (containsTag(tag, RETURN)) {
-      // @ts-ignore
-      if (tag.typeExpression) {
-        // @ts-ignore
-        type = tag.typeExpression
-          .getFullText()
-          // @ts-ignore
-          .slice(1, tag.typeExpression.getFullText().length - 1)
-          .toLowerCase();
-      }
-    }
-  });
-  return type;
-}
-
-/**
  * This method will parse out all of the @param tags of a JSDoc and return a dictionary
  * @param node - The function to parse the JSDoc params from
  */
@@ -842,7 +817,8 @@ function getJSDocParams(node: ts.Node): { [key: string]: string } {
 
   ts.getAllJSDocTagsOfKind(node, ts.SyntaxKind.JSDocParameterTag).forEach((tag: ts.JSDocTag) => {
     if (tag.comment) {
-      const comment = (tag.comment.startsWith("-") ? tag.comment.slice(1) : tag.comment).trim();
+      const tagComment = tag.comment.toString();
+      const comment = (tagComment.startsWith("-") ? tagComment.slice(1) : tagComment).trim();
       // @ts-ignore
       jsDocParamInfo[(tag as ts.JSDocPropertyLikeTag).name.getFullText()] = comment;
     } else {
@@ -859,7 +835,7 @@ function getJSDocParams(node: ts.Node): { [key: string]: string } {
  * This method will parse out all of the @param tags of a JSDoc and return a dictionary
  * @param node - The function to parse the JSDoc params from
  */
-function getJSDocParamsType(node: ts.Node): { [key: string]: string } {
+function getJSDocParamsType(node: ts.Node): { [key: string]: IJsDocParamType } {
   const jsDocParamTypeInfo = {};
 
   ts.getAllJSDocTagsOfKind(node, ts.SyntaxKind.JSDocParameterTag).forEach(
@@ -867,13 +843,35 @@ function getJSDocParamsType(node: ts.Node): { [key: string]: string } {
     (tag: ts.JSDocParameterTag) => {
       if (tag.typeExpression) {
         // Should be in the form {string}, so removing the {} around type
-        const paramType = tag.typeExpression.getFullText().slice(1, tag.typeExpression.getFullText().length - 1);
+        const paramType: string = tag.typeExpression
+          .getFullText()
+          .slice(1, tag.typeExpression.getFullText().length - 1);
+        const openBracket: number = paramType.indexOf("<");
+        let insertValue;
+
+        // Check for generic type
+        if (openBracket > -1) {
+          let subType: string = paramType.slice(openBracket + 1, paramType.length - 1);
+          const dimCount: number = (subType.match(/\[/g) || []).length;
+          if (dimCount > 0) {
+            insertValue = {
+              type: paramType.slice(0, openBracket),
+              returnType: subType.slice(0, subType.indexOf("[")),
+              dimensionality: dimCount == 1 ? "scalar" : "matrix",
+            };
+          } else {
+            insertValue = { type: paramType.slice(0, openBracket), returnType: subType };
+          }
+        } else {
+          insertValue = { type: paramType };
+        }
+
         // @ts-ignore
-        jsDocParamTypeInfo[(tag as ts.JSDocPropertyLikeTag).name.getFullText()] = paramType;
+        jsDocParamTypeInfo[(tag as ts.JSDocPropertyLikeTag).name.getFullText()] = insertValue;
       } else {
         // Set as any
         // @ts-ignore
-        jsDocParamTypeInfo[(tag as ts.JSDocPropertyLikeTag).name.getFullText()] = "any";
+        jsDocParamTypeInfo[(tag as ts.JSDocPropertyLikeTag).name.getFullText()] = { type: "any" };
       }
     }
   );
@@ -905,14 +903,14 @@ function getJSDocParamsOptionalType(node: ts.Node): { [key: string]: string } {
  */
 function hasStreamingInvocationParameter(
   param: ts.ParameterDeclaration,
-  jsDocParamTypeInfo: { [key: string]: string }
+  jsDocParamTypeInfo: { [key: string]: IJsDocParamType }
 ): boolean {
   const isTypeReferenceNode = param && param.type && ts.isTypeReferenceNode(param.type);
 
   if (param) {
     const name = (param.name as ts.Identifier).text;
-    if (name) {
-      const ptype = jsDocParamTypeInfo[name];
+    if (name && jsDocParamTypeInfo[name]) {
+      const ptype = jsDocParamTypeInfo[name].type;
       // Check to see if the streaming parameter is defined in the comment section
       if (ptype) {
         const typecheck =
@@ -945,14 +943,14 @@ function hasStreamingInvocationParameter(
  */
 function hasCancelableInvocationParameter(
   param: ts.ParameterDeclaration,
-  jsDocParamTypeInfo: { [key: string]: string }
+  jsDocParamTypeInfo: { [key: string]: IJsDocParamType }
 ): boolean {
   const isTypeReferenceNode = param && param.type && ts.isTypeReferenceNode(param.type);
 
   if (param) {
     const name = (param.name as ts.Identifier).text;
-    if (name) {
-      const ptype = jsDocParamTypeInfo[name];
+    if (name && jsDocParamTypeInfo[name]) {
+      const ptype = jsDocParamTypeInfo[name].type;
       // Check to see if the cancelable parameter is defined in the comment section
       if (ptype) {
         const cancelableTypeCheck =
@@ -981,14 +979,14 @@ function hasCancelableInvocationParameter(
  */
 function hasInvocationParameter(
   param: ts.ParameterDeclaration,
-  jsDocParamTypeInfo: { [key: string]: string }
+  jsDocParamTypeInfo: { [key: string]: IJsDocParamType }
 ): boolean {
   const isTypeReferenceNode = param && param.type && ts.isTypeReferenceNode(param.type);
 
   if (param) {
     const name = (param.name as ts.Identifier).text;
-    if (name) {
-      const ptype = jsDocParamTypeInfo[name];
+    if (name && jsDocParamTypeInfo[name]) {
+      const ptype = jsDocParamTypeInfo[name].type;
       // Check to see if the invocation parameter is defined in the comment section
       if (ptype) {
         if (ptype.toLocaleLowerCase() === TYPE_CUSTOM_FUNCTION_INVOCATION) {
@@ -1019,7 +1017,7 @@ function getParamType(t: ts.TypeNode, extra: IFunctionExtras, enumList: string[]
     if (ts.isTypeReferenceNode(t) || ts.isArrayTypeNode(t)) {
       let arrayType: IArrayType = {
         dimensionality: 0,
-        type: ts.SyntaxKind.AnyKeyword
+        type: ts.SyntaxKind.AnyKeyword,
       };
       if (ts.isTypeReferenceNode(t)) {
         const array = t as ts.TypeReferenceNode;
@@ -1051,7 +1049,7 @@ function getParamType(t: ts.TypeNode, extra: IFunctionExtras, enumList: string[]
 function getArrayDimensionalityAndType(node: ts.TypeNode): IArrayType {
   let array: IArrayType = {
     dimensionality: 0,
-    type: ts.SyntaxKind.AnyKeyword
+    type: ts.SyntaxKind.AnyKeyword,
   };
   if (ts.isArrayTypeNode(node)) {
     array = getArrayDimensionalityAndTypeForArrayTypeNode(node);
@@ -1068,7 +1066,7 @@ function getArrayDimensionalityAndType(node: ts.TypeNode): IArrayType {
 function getArrayDimensionalityAndTypeForArrayTypeNode(node: ts.TypeNode): IArrayType {
   const array: IArrayType = {
     dimensionality: 1,
-    type: ts.SyntaxKind.AnyKeyword
+    type: ts.SyntaxKind.AnyKeyword,
   };
 
   let nodeCheck = (node as ts.ArrayTypeNode).elementType;
@@ -1089,7 +1087,7 @@ function getArrayDimensionalityAndTypeForArrayTypeNode(node: ts.TypeNode): IArra
 function getArrayDimensionalityAndTypeForReferenceNode(node: ts.TypeReferenceNode): IArrayType {
   const array: IArrayType = {
     dimensionality: 0,
-    type: ts.SyntaxKind.AnyKeyword
+    type: ts.SyntaxKind.AnyKeyword,
   };
 
   if (node.typeArguments && node.typeArguments.length === 1) {
@@ -1140,14 +1138,6 @@ function getParamOptional(p: ts.ParameterDeclaration, jsDocParamOptionalInfo: { 
     optional = jsDocParamOptionalInfo[name];
   }
   return optional;
-}
-
-/**
- * This function will return `true` for `Array<[object]>` and `false` otherwise.
- * @param a - TypeReferenceNode
- */
-function validateArray(a: ts.TypeReferenceNode) {
-  return a.typeName.getText() === "Array" && a.typeArguments && a.typeArguments.length === 1;
 }
 
 /**
