@@ -5,7 +5,7 @@ import * as assert from "assert";
 import * as fs from "fs";
 import * as mocha from "mocha";
 import * as path from "path";
-import { generate } from "../../src/generate";
+import { generateCustomFunctionsMetadata } from "../../src/generate";
 
 function deleteFileIfExists(filePath: string): void {
     if (fs.existsSync(filePath)) {
@@ -38,7 +38,6 @@ describe("test cases", function() {
                     // add an expression in the file and it will be skipped if not true
                     const skip: string | undefined = readFileIfExists(path.resolve(testCaseDirPath, "skip"));
                     if (skip !== undefined) {
-                        // tslint:disable-next-line: no-eval
                         const skipResult = eval(skip);
                         if (!skipResult) {
                             this.skip();
@@ -48,24 +47,30 @@ describe("test cases", function() {
                         const expectedErrorsFile = path.join(testCaseDirPath, `expected.${scriptType}.errors.txt`);
                         const actualMetadataFile = path.join(testCaseDirPath, `actual.${scriptType}.json`);
                         const expectedMetadataFile = path.join(testCaseDirPath, "expected.json");
-                        const expectedMetadata: string | undefined = readFileIfExists(expectedMetadataFile);
+                        const expectedMetadata = readFileIfExists(expectedMetadataFile) || "";
 
                         // add a file named "debugger" to break on the test case
                         if (fs.existsSync(path.resolve(testCaseDirPath, "debugger"))) {
-                            // tslint:disable-next-line: no-debugger
                             debugger;
                         }
 
                         // generate metadata
-                        const result = await generate(sourceFile, actualMetadataFile);
+                        const result = await generateCustomFunctionsMetadata(sourceFile);
 
-                        const actualMetadata = readFileIfExists(actualMetadataFile);
+                        const actualMetadata = result.metadataJson;
                         const actualErrors = (result.errors.length > 0) ? result.errors.join("\n") : undefined;
                         const expectedErrors = readFileIfExists(expectedErrorsFile);
 
+                        if (result.errors.length > 0) {
+                            deleteFileIfExists(actualMetadataFile);
+                        } else {
+                            // write the actual metadata file
+                            fs.writeFileSync(actualMetadataFile, actualMetadata);
+                        }
+
                         // if actual errors are different than expected, write out the actual errors to a file
                         // otherwise, delete the actual errors file if it exists
-                        if (actualErrors !== expectedErrors) {
+                        if (actualErrors && actualErrors !== expectedErrors) {
                             fs.writeFileSync(actualErrorsFile, actualErrors);
                         } else {
                             deleteFileIfExists(actualErrorsFile);
