@@ -1,9 +1,4 @@
 export = {
-    extends: "eslint:recommended",
-    rules: {
-        "no-eval": "error",
-        "no-delete-var": "error",
-        "no-octal": "error",
-        "no-inner-declarations": "error"
-      }
+  plugins: ["office-addins"],
+  extends: ["plugin:office-addins/recommended"],
 };
