@@ -4,20 +4,22 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
-import { logErrorMessage } from "./log";
+import { logErrorMessage } from "office-addin-usage-data";
+
+/* global process */
 
 commander.name("office-addin-cli");
 commander.version(process.env.npm_package_version || "(version not available)");
 
 // if the command is not known, display an error
-commander.on("command:*", function() {
-    logErrorMessage(`The command syntax is not valid.\n`);
-    process.exitCode = 1;
-    commander.help();
+commander.on("command:*", function () {
+  logErrorMessage(`The command syntax is not valid.\n`);
+  process.exitCode = 1;
+  commander.help();
 });
 
 if (process.argv.length > 2) {
-    commander.parse(process.argv);
+  commander.parse(process.argv);
 } else {
-    commander.help();
+  commander.help();
 }
