@@ -4,6 +4,9 @@
 import * as childProcess from "child_process";
 import * as crypto from "crypto";
 import * as net from "net";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process */
 
 /**
  * Determines whether a port is in use.
@@ -35,10 +38,10 @@ export function isPortInUse(port: number): Promise<boolean> {
  * @example "Local Address" returns undefined
  */
 function parsePort(text: string): number | undefined {
-    const result = text.match(/:(\d+)$/);
+  const result = text.match(/:(\d+)$/);
 
-    return result ? parseInt(result[1], 10) : undefined;
-  }
+  return result ? parseInt(result[1], 10) : undefined;
+}
 
 /**
  * Return the process ids using the port.
@@ -50,7 +53,8 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
 
   return new Promise((resolve, reject) => {
     const isWin32 = process.platform === "win32";
-    const command = isWin32 ? `netstat -ano` : `lsof -n -i:${port}`;
+    const isLinux = process.platform === "linux";
+    const command = isWin32 ? `netstat -ano` : isLinux ? `netstat -tlpna | grep :${port}` : `lsof -n -i:${port}`;
 
     childProcess.exec(command, (error, stdout) => {
       if (error) {
@@ -65,6 +69,7 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
         const lines = stdout.trim().split("\n");
         if (isWin32) {
           lines.forEach((line) => {
+            /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
             const [protocol, localAddress, foreignAddress, status, processId] = line.split(" ").filter((text) => text);
             if (processId !== undefined) {
               const localAddressPort = parsePort(localAddress);
@@ -73,10 +78,26 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
               }
             }
           });
+        } else if (isLinux) {
+          lines.forEach((line) => {
+            /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
+            const [proto, recv, send, local_address, remote_address, state, program] = line
+              .split(" ")
+              .filter((text) => text);
+            if (local_address !== undefined && local_address.endsWith(`:${port}`) && program !== undefined) {
+              const pid = parseInt(program, 10);
+              if (!isNaN(pid)) {
+                processIds.add(pid);
+              }
+            }
+          });
         } else {
           lines.forEach((line) => {
-            const [process, processId, user, fd, type, device, size, node, name] = line.split(" ").filter((text) => text);
-            if ((processId !== undefined) && (processId !== "PID")) {
+            /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
+            const [process, processId, user, fd, type, device, size, node, name] = line
+              .split(" ")
+              .filter((text) => text);
+            if (processId !== undefined && processId !== "PID") {
               processIds.add(parseInt(processId, 10));
             }
           });
@@ -105,7 +126,7 @@ export async function randomPortNotInUse(): Promise<number> {
  * Returns a random number between 0 and 65535
  */
 function randomPortNumber(): number {
-   return crypto.randomBytes(2).readUInt16LE(0);
+  return crypto.randomBytes(2).readUInt16LE(0);
 }
 
 /**
@@ -114,7 +135,7 @@ function randomPortNumber(): number {
  * @throws Error if port is not a number from 0 to 65535.
  */
 function validatePort(port: number): void {
-  if ((typeof(port) !== "number") || (port < 0) || (port > 65535)) {
-    throw new Error("Port should be a number from 0 to 65535.");
+  if (typeof port !== "number" || port < 0 || port > 65535) {
+    throw new ExpectedError("Port should be a number from 0 to 65535.");
   }
 }
