@@ -1,55 +1,17 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import * as childProcess from "child_process";
-import { ChildProcess, ExecException } from "child_process";
+import { ChildProcess } from "child_process";
+import * as devSettings from "office-addin-dev-settings";
 
 export async function startProcess(commandLine: string, verbose: boolean = false): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-        if (verbose) {
-            console.log(`Starting: ${commandLine}`);
-        }
-
-        childProcess.exec(commandLine, (error: ExecException | null, stdout: string, stderr: string) => {
-            if (error) {
-                reject(error);
-            } else {
-                resolve();
-            }
-        });
-    });
+  await devSettings.startProcess(commandLine, verbose);
 }
 
 export function startDetachedProcess(commandLine: string, verbose: boolean = false): ChildProcess {
-    if (verbose) {
-        console.log(`Starting: ${commandLine}`);
-    }
-
-    const subprocess = childProcess.spawn(commandLine, [], {
-        detached: true,
-        shell: true,
-        stdio: "ignore",
-        windowsHide: false,
-    });
-
-    subprocess.on("error", (err) => {
-        console.log(`Unable to run command: ${commandLine}.\n${err}`);
-    });
-
-    subprocess.unref();
-    return subprocess;
+  return devSettings.startDetachedProcess(commandLine, verbose);
 }
 
 export function stopProcess(processId: number): void {
-    if (processId) {
-        try {
-            if (process.platform === "win32") {
-                childProcess.spawn("taskkill", ["/pid", `${processId}`, "/f", "/t"]);
-            } else {
-                process.kill(processId);
-            }
-        } catch (err) {
-            console.log(`Unable to kill process id ${processId}: ${err}`);
-        }
-    }
+  devSettings.stopProcess(processId);
 }
