@@ -3,10 +3,21 @@
 
 import * as fs from "fs-extra";
 import * as junk from "junk";
-import { getOfficeApps, getOfficeAppsForManifestHosts, OfficeApp, readManifestFile } from "office-addin-manifest";
+import {
+  exportMetadataPackage,
+  getOfficeApps,
+  getOfficeAppsForManifestHosts,
+  OfficeApp,
+  OfficeAddinManifest,
+} from "office-addin-manifest";
 import * as os from "os";
 import * as path from "path";
 import { RegisteredAddin } from "./dev-settings";
+import { ExpectedError } from "office-addin-usage-data";
+import { registerWithTeams } from "./publish";
+import * as fspath from "path";
+
+/* global process */
 
 export async function getRegisteredAddIns(): Promise<RegisteredAddin[]> {
   const registeredAddins: RegisteredAddin[] = [];
@@ -17,7 +28,7 @@ export async function getRegisteredAddIns(): Promise<RegisteredAddin[]> {
     if (sideloadDirectory && fs.existsSync(sideloadDirectory)) {
       for (const fileName of fs.readdirSync(sideloadDirectory).filter(junk.not)) {
         const manifestPath = fs.realpathSync(path.join(sideloadDirectory, fileName));
-        const manifest = await readManifestFile(manifestPath);
+        const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
         registeredAddins.push(new RegisteredAddin(manifest.id || "", manifestPath));
       }
     }
@@ -39,29 +50,35 @@ function getSideloadDirectory(app: OfficeApp): string | undefined {
 
 export async function registerAddIn(manifestPath: string, officeApps?: OfficeApp[]) {
   try {
-    const manifest = await readManifestFile(manifestPath);
+    const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
 
     if (!officeApps) {
       officeApps = getOfficeAppsForManifestHosts(manifest.hosts);
 
       if (officeApps.length === 0) {
-        throw new Error("The manifest file doesn't specify any hosts for the Office Add-in.");
+        throw new ExpectedError("The manifest file doesn't specify any hosts for the Office Add-in.");
       }
     }
 
     if (!manifest.id) {
-      throw new Error("The manifest file doesn't contain the id of the Office Add-in.");
+      throw new ExpectedError("The manifest file doesn't contain the id of the Office Add-in.");
     }
 
-    for (const app of officeApps) {
-      const sideloadDirectory = getSideloadDirectory(app);
+    if (manifestPath.endsWith(".json")) {
+      const targetPath: string = fspath.join(process.env.TEMP as string, "manifest.zip");
+      const zipPath: string = await exportMetadataPackage(targetPath, manifestPath);
+      return registerWithTeams(zipPath);
+    } else if (manifestPath.endsWith(".xml")) {
+      for (const app of officeApps) {
+        const sideloadDirectory = getSideloadDirectory(app);
 
-      if (sideloadDirectory) {
-        // include manifest id in sideload filename
-        const sideloadPath = path.join(sideloadDirectory, `${manifest.id}.${path.basename(manifestPath)}`);
+        if (sideloadDirectory) {
+          // include manifest id in sideload filename
+          const sideloadPath = path.join(sideloadDirectory, `${manifest.id}.${path.basename(manifestPath)}`);
 
-        fs.ensureDirSync(sideloadDirectory);
-        fs.ensureLinkSync(manifestPath, sideloadPath);
+          fs.ensureDirSync(sideloadDirectory);
+          fs.ensureLinkSync(manifestPath, sideloadPath);
+        }
       }
     }
   } catch (err) {
@@ -70,10 +87,10 @@ export async function registerAddIn(manifestPath: string, officeApps?: OfficeApp
 }
 
 export async function unregisterAddIn(manifestPath: string): Promise<void> {
-  const manifest = await readManifestFile(manifestPath);
+  const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
 
   if (!manifest.id) {
-    throw new Error("The manifest file doesn't contain the id of the Office Add-in.");
+    throw new ExpectedError("The manifest file doesn't contain the id of the Office Add-in.");
   }
 
   const registeredAddIns = await getRegisteredAddIns();
@@ -82,8 +99,7 @@ export async function unregisterAddIn(manifestPath: string): Promise<void> {
     const registeredFileName = path.basename(registeredAddIn.manifestPath);
     const manifestFileName = path.basename(manifestPath);
     const sideloadFileName = `${manifest.id!}.${manifestFileName}`;
-    if ((registeredFileName === manifestFileName)
-      || (registeredFileName === sideloadFileName)) {
+    if (registeredFileName === manifestFileName || registeredFileName === sideloadFileName) {
       fs.unlinkSync(registeredAddIn.manifestPath);
     }
   }
