@@ -2,13 +2,17 @@
 // Licensed under the MIT license.
 
 import * as fs from "fs";
-import { readManifestFile } from "office-addin-manifest";
+import { OfficeAddinManifest } from "office-addin-manifest";
 import * as fspath from "path";
 import * as devSettingsMac from "./dev-settings-mac";
 import * as devSettingsWindows from "./dev-settings-windows";
-import { platform } from "os";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process */
 
 const defaultRuntimeLogFileName = "OfficeAddins.log.txt";
+
+export { toWebViewTypeName } from "./dev-settings-windows";
 
 export enum DebuggingMethod {
   Direct,
@@ -41,10 +45,10 @@ export class SourceBundleUrlComponents {
   public extension?: string;
 
   public get url(): string {
-    const host = (this.host !== undefined) ? this.host : "localhost";
-    const port = (this.port !== undefined) ? this.port : "8081";
-    const path = (this.path !== undefined) ? this.path : "{path}";
-    const extension = (this.extension !== undefined) ? this.extension : ".bundle";
+    const host = this.host !== undefined ? this.host : "localhost";
+    const port = this.port !== undefined ? this.port : "8081";
+    const path = this.path !== undefined ? this.path : "{path}";
+    const extension = this.extension !== undefined ? this.extension : ".bundle";
 
     return `http://${host}${host && port ? ":" : ""}${port}/${path}${extension}`;
   }
@@ -61,8 +65,8 @@ export async function clearDevSettings(addinId: string): Promise<void> {
   switch (process.platform) {
     case "win32":
       return devSettingsWindows.clearDevSettings(addinId);
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -78,17 +82,22 @@ export async function disableRuntimeLogging(): Promise<void> {
   switch (process.platform) {
     case "win32":
       return devSettingsWindows.disableRuntimeLogging();
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
-export async function enableDebugging(addinId: string, enable: boolean = true, method: DebuggingMethod = DebuggingMethod.Direct): Promise<void> {
+export async function enableDebugging(
+  addinId: string,
+  enable: boolean = true,
+  method: DebuggingMethod = DebuggingMethod.Direct,
+  openDevTools = false
+): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return devSettingsWindows.enableDebugging(addinId, enable, method);
+      return devSettingsWindows.enableDebugging(addinId, enable, method, openDevTools);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -96,18 +105,18 @@ export async function enableLiveReload(addinId: string, enable: boolean = true):
   switch (process.platform) {
     case "win32":
       return devSettingsWindows.enableLiveReload(addinId, enable);
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function enableRuntimeLogging(path?: string): Promise<string> {
   switch (process.platform) {
-    case "win32":
+    case "win32": {
       if (!path) {
         const tempDir = process.env.TEMP;
         if (!tempDir) {
-          throw new Error("The TEMP environment variable is not defined.");
+          throw new ExpectedError("The TEMP environment variable is not defined.");
         }
         path = fspath.normalize(`${tempDir}/${defaultRuntimeLogFileName}`);
       }
@@ -116,36 +125,39 @@ export async function enableRuntimeLogging(path?: string): Promise<string> {
       if (pathExists) {
         const stat = fs.statSync(path);
         if (stat.isDirectory()) {
-          throw new Error(`You need to specify the path to a file. This is a directory: "${path}".`);
+          throw new ExpectedError(`You need to specify the path to a file. This is a directory: "${path}".`);
         }
       }
       try {
         const file = fs.openSync(path, "a+");
         fs.closeSync(file);
       } catch (err) {
-        throw new Error(pathExists
-          ? `You need to specify the path to a writable file. Unable to write to: "${path}".`
-          : `You need to specify the path where the file can be written. Unable to write to: "${path}".`);
+        throw new ExpectedError(
+          pathExists
+            ? `You need to specify the path to a writable file. Unable to write to: "${path}".`
+            : `You need to specify the path where the file can be written. Unable to write to: "${path}".`
+        );
       }
 
       await devSettingsWindows.enableRuntimeLogging(path);
       return path;
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    }
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 /**
  * Returns the manifest paths for the add-ins that are registered
  */
-export async function getRegisterAddIns(): Promise<RegisteredAddin[]>  {
+export async function getRegisterAddIns(): Promise<RegisteredAddin[]> {
   switch (process.platform) {
     case "darwin":
       return devSettingsMac.getRegisteredAddIns();
     case "win32":
       return devSettingsWindows.getRegisteredAddIns();
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -154,7 +166,16 @@ export async function getEnabledDebuggingMethods(addinId: string): Promise<Debug
     case "win32":
       return devSettingsWindows.getEnabledDebuggingMethods(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
+
+export async function getOpenDevTools(addinId: string): Promise<boolean> {
+  switch (process.platform) {
+    case "win32":
+      return devSettingsWindows.getOpenDevTools(addinId);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -163,7 +184,7 @@ export async function getRuntimeLoggingPath(): Promise<string | undefined> {
     case "win32":
       return devSettingsWindows.getRuntimeLoggingPath();
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -172,7 +193,7 @@ export async function getSourceBundleUrl(addinId: string): Promise<SourceBundleU
     case "win32":
       return devSettingsWindows.getSourceBundleUrl(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -181,7 +202,7 @@ export async function getWebView(addinId: string): Promise<WebViewType | undefin
     case "win32":
       return devSettingsWindows.getWebView(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -190,7 +211,7 @@ export async function isDebuggingEnabled(addinId: string): Promise<boolean> {
     case "win32":
       return devSettingsWindows.isDebuggingEnabled(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -199,20 +220,20 @@ export async function isLiveReloadEnabled(addinId: string): Promise<boolean> {
     case "win32":
       return devSettingsWindows.isLiveReloadEnabled(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function registerAddIn(manifestPath: string): Promise<void> {
   switch (process.platform) {
-    case "win32":
-      const manifest = await readManifestFile(manifestPath);
+    case "win32": {
       const realManifestPath = fs.realpathSync(manifestPath);
-      return devSettingsWindows.registerAddIn(manifest.id || "", realManifestPath);
+      return devSettingsWindows.registerAddIn(realManifestPath);
+    }
     case "darwin":
       return devSettingsMac.registerAddIn(manifestPath);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -221,16 +242,16 @@ export async function setSourceBundleUrl(addinId: string, components: SourceBund
     case "win32":
       return devSettingsWindows.setSourceBundleUrl(addinId, components);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function setWebView(addinId: string, webViewType: WebViewType | undefined) {
   switch (process.platform) {
     case "win32":
-      return devSettingsWindows.setWebView(addinId, webViewType);;
+      return devSettingsWindows.setWebView(addinId, webViewType);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -238,12 +259,13 @@ export async function unregisterAddIn(manifestPath: string): Promise<void> {
   switch (process.platform) {
     case "darwin":
       return devSettingsMac.unregisterAddIn(manifestPath);
-    case "win32":
-      const manifest = await readManifestFile(manifestPath);
+    case "win32": {
+      const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
       const realManifestPath = fs.realpathSync(manifestPath);
       return devSettingsWindows.unregisterAddIn(manifest.id || "", realManifestPath);
+    }
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -254,6 +276,6 @@ export async function unregisterAllAddIns(): Promise<void> {
     case "win32":
       return devSettingsWindows.unregisterAllAddIns();
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
