@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 import * as fs from "fs";
+import * as childProcess from "child_process";
 import * as jszip from "jszip";
 import {
   AddInType,
@@ -10,14 +11,22 @@ import {
   getOfficeAppsForManifestHosts,
   ManifestInfo,
   OfficeApp,
-  readManifestFile,
+  OfficeAddinManifest,
 } from "office-addin-manifest";
 import open = require("open");
+import semver = require("semver");
 import * as os from "os";
 import * as path from "path";
 import * as util from "util";
+import { AppType } from "./appType";
 import { registerAddIn } from "./dev-settings";
+import { startDetachedProcess } from "./process";
 import { chooseOfficeApp } from "./prompt";
+import * as registry from "./registry";
+import { usageDataObject } from "./defaults";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process, __dirname, URL, console */
 
 const readFileAsync = util.promisify(fs.readFile);
 
@@ -28,53 +37,57 @@ const readFileAsync = util.promisify(fs.readFile);
  * @param manifest Manifest for the add-in.
  * @returns Path to the file.
  */
-async function generateSideloadFile(app: OfficeApp, manifest: ManifestInfo): Promise<string> {
+export async function generateSideloadFile(app: OfficeApp, manifest: ManifestInfo, document?: string): Promise<string> {
   if (!manifest.id) {
-    throw new Error("The manifest does not contain the id for the add-in.");
+    throw new ExpectedError("The manifest does not contain the id for the add-in.");
   }
 
   if (!manifest.officeAppType) {
-    throw new Error("The manifest does not contain the OfficeApp xsi:type.");
+    throw new ExpectedError("The manifest does not contain the OfficeApp xsi:type.");
   }
 
   if (!manifest.version) {
-    throw new Error(
-      "The manifest does not contain the version for the add-in.",
-    );
+    throw new ExpectedError("The manifest does not contain the version for the add-in.");
   }
-
   const addInType = getAddInTypeForManifestOfficeAppType(manifest.officeAppType);
 
   if (!addInType) {
-    throw new Error("The manifest contains an unsupported OfficeApp xsi:type.");
+    throw new ExpectedError("The manifest contains an unsupported OfficeApp xsi:type.");
   }
 
-  const templatePath = getTemplatePath(app, addInType);
+  const documentWasProvided = document && document !== "";
+  const templatePath = documentWasProvided ? path.resolve(document) : getTemplatePath(app, addInType);
 
   if (!templatePath) {
-    throw new Error("Sideload is not supported.");
+    throw new ExpectedError(`Sideload is not supported for apptype: ${addInType}.`);
   }
 
   const templateBuffer = await readFileAsync(templatePath);
-  const zip = await jszip.loadAsync(templateBuffer);
-  const webExtensionPath = getWebExtensionPath(app, addInType);
-
-  if (!webExtensionPath) {
-    throw new Error("Don't know the webextension path.");
-  }
+  const zip: jszip = await jszip.loadAsync(templateBuffer);
 
   const appName = getOfficeAppName(app);
   const extension = path.extname(templatePath);
-  const pathToWrite = makePathUnique(
+  const pathToWrite: string = makePathUnique(
     path.join(os.tmpdir(), `${appName} add-in ${manifest.id}${extension}`),
-    true,
+    true
   );
 
-  // replace the placeholder id and version
-  const webExtensionXml = (await zip.file(webExtensionPath).async("text"))
-    .replace(/00000000-0000-0000-0000-000000000000/g, manifest.id)
-    .replace(/1.0.0.0/g, manifest.version);
-  zip.file(webExtensionPath, webExtensionXml);
+  if (!documentWasProvided) {
+    const webExtensionPath = getWebExtensionPath(app, addInType);
+    if (!webExtensionPath) {
+      throw new ExpectedError("Don't know the webextension path.");
+    }
+
+    // replace the placeholder id and version
+    const zipFile = zip.file(webExtensionPath);
+    if (!zipFile) {
+      throw new ExpectedError("webextension was not found.");
+    }
+    const webExtensionXml = (await zipFile.async("text"))
+      .replace(/00000000-0000-0000-0000-000000000000/g, manifest.id)
+      .replace(/1.0.0.0/g, manifest.version);
+    zip.file(webExtensionPath, webExtensionXml);
+  }
 
   // Write the file
   await new Promise((resolve, reject) => {
@@ -89,15 +102,60 @@ async function generateSideloadFile(app: OfficeApp, manifest: ManifestInfo): Pro
 }
 
 /**
+ * Create an Office document url with query params which can be opened
+ * to register an Office add-in in Office Online.
+ * @param manifestPath Path to the manifest file for the Office Add-in.
+ * @param documentUrl Office Online document url
+ * @param isTest Indicates whether to append test query param to suppress Office Online dialogs.
+ * @returns Document url with query params appended.
+ */
+export async function generateSideloadUrl(
+  manifestFileName: string,
+  manifest: ManifestInfo,
+  documentUrl: string | undefined,
+  isTest: boolean = false
+): Promise<string | undefined> {
+  const testQueryParam = "&wdaddintest=true";
+
+  if (documentUrl === undefined || documentUrl === "") {
+    return undefined;
+  }
+
+  if (!manifest.id) {
+    throw new ExpectedError("The manifest does not contain the id for the add-in.");
+  }
+
+  if (manifest.defaultSettings === undefined || manifest.defaultSettings.sourceLocation === undefined) {
+    throw new ExpectedError("The manifest does not contain the SourceLocation for the add-in");
+  }
+
+  const sourceLocationUrl: URL = new URL(manifest.defaultSettings.sourceLocation);
+  if (sourceLocationUrl.protocol.indexOf("https") === -1) {
+    throw new ExpectedError("The SourceLocation in the manifest does not use the HTTPS protocol.");
+  }
+
+  if (sourceLocationUrl.host.indexOf("localhost") === -1 && sourceLocationUrl.host.indexOf("127.0.0.1") === -1) {
+    throw new ExpectedError(
+      "The hostname specified by the SourceLocation in the manifest is not supported for sideload. The hostname should be 'localhost' or 127.0.0.1."
+    );
+  }
+
+  let queryParms: string = `&wdaddindevserverport=${sourceLocationUrl.port}&wdaddinmanifestfile=${manifestFileName}&wdaddinmanifestguid=${manifest.id}`;
+
+  if (isTest) {
+    queryParms = `${queryParms}${testQueryParam}`;
+  }
+
+  return `${documentUrl}${queryParms}`;
+}
+
+/**
  * Returns the path to the document used as a template for sideloading,
  * or undefined if sideloading is not supported.
  * @param app Specifies the Office app.
  * @param addInType Specifies the type of add-in.
  */
-function getTemplatePath(
-  app: OfficeApp,
-  addInType: AddInType,
-): string | undefined {
+export function getTemplatePath(app: OfficeApp, addInType: AddInType): string | undefined {
   switch (app) {
     case OfficeApp.Excel:
       switch (addInType) {
@@ -129,10 +187,7 @@ function getTemplatePath(
  * @param app Specifies the Office app.
  * @param addInType Specifies the type of add-in.
  */
-function getWebExtensionPath(
-  app: OfficeApp,
-  addInType: AddInType,
-): string | undefined {
+function getWebExtensionPath(app: OfficeApp, addInType: AddInType): string | undefined {
   switch (app) {
     case OfficeApp.Excel:
       return "xl/webextensions/webextension.xml";
@@ -146,6 +201,54 @@ function getWebExtensionPath(
       break;
     case OfficeApp.Word:
       return "word/webextensions/webextension.xml";
+  }
+}
+
+function isSideloadingSupportedForDesktopHost(app: OfficeApp): boolean {
+  if (
+    app === OfficeApp.Excel ||
+    (app === OfficeApp.Outlook && process.platform === "win32") ||
+    app === OfficeApp.PowerPoint ||
+    app === OfficeApp.Word
+  ) {
+    return true;
+  }
+  return false;
+}
+
+function isSideloadingSupportedForWebHost(app: OfficeApp): boolean {
+  if (app === OfficeApp.Excel || app === OfficeApp.PowerPoint || app === OfficeApp.Project || app === OfficeApp.Word) {
+    return true;
+  }
+  return false;
+}
+
+function hasOfficeVersion(targetVersion: string, currentVersion: string): boolean {
+  return semver.gte(currentVersion, targetVersion);
+}
+
+async function getOutlookVersion(): Promise<string | undefined> {
+  try {
+    const key = new registry.RegistryKey(`HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Office\\ClickToRun\\Configuration`);
+    const outlookInstallVersion: string | undefined = await registry.getStringValue(key, "ClientVersionToReport");
+    const outlookSmallerVersion = outlookInstallVersion?.split(`.`, 3).join(`.`);
+
+    return outlookSmallerVersion;
+  } catch (err) {
+    return undefined;
+  }
+}
+
+async function getOutlookExePath(): Promise<string | undefined> {
+  try {
+    const OutlookInstallPathRegistryKey: string = `HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\App Paths\\OUTLOOK.EXE`;
+    const key = new registry.RegistryKey(`${OutlookInstallPathRegistryKey}`);
+    const outlookExePath: string | undefined = await registry.getStringValue(key, "");
+
+    return outlookExePath;
+  } catch (err) {
+    const errorMessage: string = `Unable to find Outlook install location: \n${err}`;
+    throw new Error(errorMessage);
   }
 }
 
@@ -178,10 +281,7 @@ function makePathUnique(originalPath: string, tryToDelete: boolean = false): str
         parsedPath = path.parse(originalPath);
       }
 
-      currentPath = path.join(
-        parsedPath.dir,
-        `${parsedPath.name}.${suffix}${parsedPath.ext}`,
-      );
+      currentPath = path.join(parsedPath.dir, `${parsedPath.name}.${suffix}${parsedPath.ext}`);
     }
   }
 
@@ -194,37 +294,169 @@ function makePathUnique(originalPath: string, tryToDelete: boolean = false): str
  * @param app Office app to launch.
  * @param canPrompt
  */
-export async function sideloadAddIn(manifestPath: string, app?: OfficeApp, canPrompt: boolean = false): Promise<void> {
-  const manifest = await readManifestFile(manifestPath);
-
-  const appsInManifest = getOfficeAppsForManifestHosts(manifest.hosts);
-
-  if (app) {
-    if (appsInManifest.indexOf(app) < 0) {
-      throw new Error(`The Office Add-in manifest does not support ${getOfficeAppName(app)}.`);
+export async function sideloadAddIn(
+  manifestPath: string,
+  app?: OfficeApp,
+  canPrompt: boolean = false,
+  appType?: AppType,
+  document?: string
+): Promise<void> {
+  try {
+    if (appType === undefined) {
+      appType = AppType.Desktop;
     }
-  } else {
-    switch (appsInManifest.length) {
-      case 0:
-        throw new Error("The manifest does not support any Office apps.");
-      case 1:
-        app = appsInManifest[0];
-        break;
-      default:
-        if (canPrompt) {
-          app = await chooseOfficeApp(appsInManifest);
+
+    // Converting json to xml manifest . . . Temporary until service is ready.
+    if (manifestPath.endsWith(".json")) {
+      if (isDotnetInstalled()) {
+        // Run json => xml conversion tool.
+        manifestPath = await convertJsonToXmlManifest(manifestPath);
+      } else {
+        throw new ExpectedError(".Net 5 or greater is required for json manifests.");
+      }
+    }
+
+    const manifest: ManifestInfo = await OfficeAddinManifest.readManifestFile(manifestPath);
+    const appsInManifest: OfficeApp[] = getOfficeAppsForManifestHosts(manifest.hosts);
+    const isTest: boolean = process.env.WEB_SIDELOAD_TEST !== undefined;
+
+    if (app) {
+      if (appsInManifest.indexOf(app) < 0) {
+        throw new ExpectedError(`The Office Add-in manifest does not support ${getOfficeAppName(app)}.`);
+      }
+    } else {
+      switch (appsInManifest.length) {
+        case 0:
+          throw new ExpectedError("The manifest does not support any Office apps.");
+        case 1:
+          app = appsInManifest[0];
+          break;
+        default:
+          if (canPrompt) {
+            app = await chooseOfficeApp(appsInManifest);
+          }
+          break;
+      }
+    }
+
+    if (!app) {
+      throw new ExpectedError("Please specify the Office app.");
+    }
+
+    let sideloadFile: string | undefined;
+
+    switch (appType) {
+      case AppType.Desktop:
+        if (!isSideloadingSupportedForDesktopHost(app)) {
+          throw new ExpectedError(`Sideload to the ${getOfficeAppName(app)} app is not supported.`);
+        }
+
+        // do the registration
+        await registerAddIn(manifestPath);
+        // for Outlook, open Outlook.exe; for other Office apps, open the document
+        if (app == OfficeApp.Outlook) {
+          const version: string | undefined = await getOutlookVersion();
+          if (version && !hasOfficeVersion("16.0.13709", version)) {
+            throw new ExpectedError(
+              `The current version of Outlook does not support sideload. Please use version 16.0.13709 or greater.`
+            );
+          }
+          sideloadFile = await getOutlookExePath();
         } else {
-          throw new Error("Please specify the Office app.");
+          sideloadFile = await generateSideloadFile(app, manifest, document);
         }
         break;
+      case AppType.Web: {
+        if (!document) {
+          throw new ExpectedError(`For sideload to web, you need to specify a document url.`);
+        }
+        if (!isSideloadingSupportedForWebHost(app)) {
+          throw new ExpectedError(`Sideload to the ${getOfficeAppName(app)} web app is not supported.`);
+        }
+        const manifestFileName: string = path.basename(manifestPath);
+        sideloadFile = await generateSideloadUrl(manifestFileName, manifest, document, isTest);
+        break;
+      }
+      default:
+        throw new ExpectedError("Sideload is not supported for the specified app type.");
     }
+
+    if (sideloadFile) {
+      if (app === OfficeApp.Outlook) {
+        // put the Outlook.exe path in quotes if it contains spaces
+        if (sideloadFile.indexOf(" ") >= 0) {
+          sideloadFile = `"${sideloadFile}"`;
+        }
+
+        startDetachedProcess(sideloadFile);
+      } else {
+        await open(sideloadFile, { wait: false });
+      }
+    }
+    usageDataObject.reportSuccess("sideloadAddIn()");
+  } catch (err: any) {
+    usageDataObject.reportException("sideloadAddIn()", err);
+    throw err;
   }
+}
 
-  await registerAddIn(manifestPath);
+function isDotnetInstalled(): boolean {
+  try {
+    // Find the .Net runtimes installed
+    let result = childProcess.execSync("dotnet --list-runtimes");
+    const pattern = /(?<=Microsoft.NETCore.App )[\d.]+/g;
+    const matches = result.toString("utf-8").match(pattern);
+    let foundDotNet = false;
 
-  if (app) {
-    const sideloadFile = await generateSideloadFile(app, manifest);
+    // Look for version 5 or greater
+    matches?.forEach((match) => {
+      const major: number = parseInt(match.split(".")[0]);
+      foundDotNet = foundDotNet || major >= 5;
+    });
 
-    await open(sideloadFile, { wait: false });
+    return foundDotNet;
+  } catch (err) {
+    return false;
+  }
+}
+
+async function convertJsonToXmlManifest(manifestPath: string): Promise<string> {
+  return new Promise((resolve, reject) => {
+    if (manifestPath.endsWith(".json") && fs.existsSync(manifestPath)) {
+      console.log("Converting json to back compat xml");
+      const convertToolPath = path.resolve(__dirname, ".\\DevXTool.exe");
+      const newManifestPath = path.join(process.env.TEMP as string, "manifest.xml");
+      const command = `"${convertToolPath}" "${manifestPath}" "${newManifestPath}"`;
+
+      childProcess.exec(command, (error, stdout) => {
+        if (error) {
+          console.log(`Error converting file:\n ${stdout}\n ${error}`);
+          reject("");
+        } else {
+          console.log(`Successfully converted manifest to xml:\n ${stdout}`);
+          stripBom(newManifestPath);
+          resolve(newManifestPath);
+        }
+      });
+    } else {
+      reject(new Error(`The file '${manifestPath}' is not valid`));
+    }
+  });
+}
+
+export default function stripBom(manifestPath: string) {
+  try {
+    if (fs.existsSync(manifestPath)) {
+      const fileData: string = fs.readFileSync(manifestPath, "utf8");
+      let updateFileData: string;
+      if (fileData.charCodeAt(0) === 0xfeff) {
+        updateFileData = fileData.slice(1);
+      } else {
+        updateFileData = fileData;
+      }
+      fs.writeFileSync(manifestPath, updateFileData);
+    }
+  } catch (err) {
+    console.log("Error trying to update converted xml");
   }
 }
