@@ -5,9 +5,21 @@ import * as usageData from "office-addin-usage-data";
 
 export const lintFiles = "src/**/*.{ts,tsx,js,jsx}";
 
+export enum ESLintExitCode {
+  NoLintErrors = 0,
+  HasLintError = 1,
+  CommandFailed = 2,
+}
+
+export enum PrettierExitCode {
+  NoFormattingProblems = 0,
+  HasFormattingProblem = 1,
+  CommandFailed = 2,
+}
+
 // Usage data defaults
 export const usageDataObject: usageData.OfficeAddinUsageData = new usageData.OfficeAddinUsageData({
-    projectName: "office-addin-lint",
-    instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
-    raisePrompt: false
+  projectName: "office-addin-lint",
+  instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
+  raisePrompt: false,
 });
