@@ -3,30 +3,34 @@
 
 import chalk from "chalk";
 import * as commander from "commander";
-import { logErrorMessage } from "office-addin-cli";
-import * as manifestInfo from "./manifestInfo";
+import { logErrorMessage } from "office-addin-usage-data";
+import { ManifestInfo } from "./manifestInfo";
+import { OfficeAddinManifest } from "./manifestOperations";
 import { ManifestValidation, ManifestValidationIssue, ManifestValidationProduct, validateManifest } from "./validate";
-import { usageDataObject } from './defaults';
+import { usageDataObject } from "./defaults";
+import { exportMetadataPackage } from "./export";
+
+/* global console, process */
 
 function getCommandOptionString(option: string | boolean, defaultValue?: string): string | undefined {
   // For a command option defined with an optional value, e.g. "--option [value]",
   // when the option is provided with a value, it will be of type "string", return the specified value;
   // when the option is provided without a value, it will be of type "boolean", return undefined.
-  return (typeof(option) === "boolean") ? defaultValue : option;
+  return typeof option === "boolean" ? defaultValue : option;
 }
 
 export async function info(manifestPath: string) {
   try {
-    const manifest = await manifestInfo.readManifestFile(manifestPath);
+    const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
     logManifestInfo(manifestPath, manifest);
-    usageDataObject.sendUsageDataSuccessEvent("info");
-  } catch (err) {
-    usageDataObject.sendUsageDataException("info", err);
+    usageDataObject.reportSuccess("info");
+  } catch (err: any) {
+    usageDataObject.reportException("info", err);
     logErrorMessage(err);
   }
 }
 
-function logManifestInfo(manifestPath: string, manifest: manifestInfo.ManifestInfo) {
+function logManifestInfo(manifestPath: string, manifest: ManifestInfo) {
   console.log(`Manifest: ${manifestPath}`);
   console.log(`  Id: ${manifest.id || ""}`);
   console.log(`  Name: ${manifest.displayName || ""}`);
@@ -57,8 +61,9 @@ function logManifestValidationErrors(errors: ManifestValidationIssue[] | undefin
   if (errors) {
     let errorNumber = 1;
     for (const currentError of errors) {
-      console.log(chalk.bold.red(`\nError # ${errorNumber}: `));
+      console.log(chalk.bold.red(`Error #${errorNumber}: `));
       logManifestValidationIssue(currentError);
+      console.log();
       ++errorNumber;
     }
   }
@@ -66,7 +71,7 @@ function logManifestValidationErrors(errors: ManifestValidationIssue[] | undefin
 
 function logManifestValidationInfos(infos: ManifestValidationIssue[] | undefined) {
   if (infos) {
-    console.log(chalk.bold.blue(`\nAdditional information: `));
+    console.log(chalk.bold.blue(`Validation Information: `));
     for (const currentInfo of infos) {
       logManifestValidationIssue(currentInfo);
       console.log();
@@ -78,8 +83,9 @@ function logManifestValidationWarnings(warnings: ManifestValidationIssue[] | und
   if (warnings) {
     let warningNumber = 1;
     for (const currentWarning of warnings) {
-      console.log(chalk.bold.yellow(`\nWarning # ${warningNumber}: `));
+      console.log(chalk.bold.yellow(`Warning #${warningNumber}: `));
       logManifestValidationIssue(currentWarning);
+      console.log();
       ++warningNumber;
     }
   }
@@ -101,15 +107,21 @@ function logManifestValidationIssue(issue: ManifestValidationIssue) {
 
 function logManifestValidationSupportedProducts(products: ManifestValidationProduct[] | undefined) {
   if (products) {
-    const productTitles = new Set(products.filter(product => product.title).map(product => product.title));
+    const productTitles = new Set(products.filter((product) => product.title).map((product) => product.title));
 
     if (productTitles.size > 0) {
-      console.log(`\nBased on the requirements specified in your manifest, your add-in can run on the following platforms; your add-in will be tested on these platforms when you submit it to the Office Store:`);
+      console.log(
+        `\nBased on the requirements specified in your manifest, your add-in can run on the following platforms; your add-in will be tested on these platforms when you submit it to the Office Store:`
+      );
       for (const productTitle of productTitles) {
         console.log(`  - ${productTitle}`);
       }
-      console.log(`Important: This analysis is based on the requirements specified in your manifest and does not account for any runtime JavaScript calls within your add-in. For information about which API sets and features are supported on each platform, see Office Add-in host and platform availability. (https://docs.microsoft.com/office/dev/add-ins/overview/office-add-in-availability).\n`);
-      console.log(`*This does not include mobile apps. You can opt-in to support mobile apps when you submit your add-in.`);
+      console.log(
+        `Important: This analysis is based on the requirements specified in your manifest and does not account for any runtime JavaScript calls within your add-in. For information about which API sets and features are supported on each platform, see Office Add-in host and platform availability. (https://docs.microsoft.com/office/dev/add-ins/overview/office-add-in-availability).\n`
+      );
+      console.log(
+        `*This does not include mobile apps. You can opt-in to support mobile apps when you submit your add-in.`
+      );
     }
   }
 }
@@ -120,36 +132,55 @@ export async function modify(manifestPath: string, command: commander.Command) {
     const guid: string | undefined = getCommandOptionString(command.guid, "");
     const displayName: string | undefined = getCommandOptionString(command.displayName);
 
-    const manifest = await manifestInfo.modifyManifestFile(manifestPath, guid, displayName);
+    const manifest = await OfficeAddinManifest.modifyManifestFile(manifestPath, guid, displayName);
     logManifestInfo(manifestPath, manifest);
-  } catch (err) {
+    usageDataObject.reportSuccess("modify");
+  } catch (err: any) {
+    usageDataObject.reportException("modify", err);
     logErrorMessage(err);
   }
 }
 
-export async function validate(manifestPath: string, command: commander.Command) {
+export async function validate(
+  manifestPath: string,
+  command: commander.Command /* eslint-disable-line @typescript-eslint/no-unused-vars */
+) {
   try {
-    const validation: ManifestValidation = await validateManifest(manifestPath);
-
-    if (validation.isValid) {
-      console.log("The manifest is valid.");
-    } else {
-      console.log("The manifest is not valid.");
-    }
-    console.log();
-
+    const verifyProduction: boolean = command.production;
+    const validation: ManifestValidation = await validateManifest(manifestPath, verifyProduction);
     if (validation.report) {
+      logManifestValidationInfos(validation.report.notes);
       logManifestValidationErrors(validation.report.errors);
       logManifestValidationWarnings(validation.report.warnings);
-      logManifestValidationInfos(validation.report.notes);
 
       if (validation.isValid) {
         if (validation.report.addInDetails) {
           logManifestValidationSupportedProducts(validation.report.addInDetails.supportedProducts);
+          console.log();
         }
+        console.log(chalk.bold.green("The manifest is valid.\n"));
+      } else {
+        console.log(chalk.bold.red("The manifest is not valid.\n"));
       }
     }
-  } catch (err) {
+
+    process.exitCode = validation.isValid ? 0 : 1;
+    usageDataObject.reportSuccess("validate");
+  } catch (err: any) {
+    usageDataObject.reportException("validate", err);
+    logErrorMessage(err);
+  }
+}
+
+export async function exportManifest(command: commander.Command) {
+  try {
+    const outputPath: string = command.output ?? "";
+    const manifestPath: string = command.manifest ?? "./manifest.json";
+
+    await exportMetadataPackage(outputPath, manifestPath);
+    usageDataObject.reportSuccess("export");
+  } catch (err: any) {
+    usageDataObject.reportException("export", err);
     logErrorMessage(err);
   }
 }
