@@ -5,7 +5,7 @@ import * as usageData from "office-addin-usage-data";
 
 // Usage data defaults
 export const usageDataObject: usageData.OfficeAddinUsageData = new usageData.OfficeAddinUsageData({
-    projectName: "office-addin-manifest",
-    instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
-    raisePrompt: false
+  projectName: "office-addin-manifest",
+  instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
+  raisePrompt: false,
 });
