@@ -17,7 +17,7 @@ export enum OfficeApp {
 }
 
 // initialized once since this list won't change
-const officeApps: OfficeApp[] = Object.keys(OfficeApp).map<OfficeApp>(key => parseOfficeApp(key));
+const officeApps: OfficeApp[] = Object.keys(OfficeApp).map<OfficeApp>((key) => parseOfficeApp(key));
 
 /**
  * Get the Office app for the manifest Host name
@@ -27,6 +27,8 @@ export function getOfficeAppForManifestHost(host: string): OfficeApp | undefined
   switch (host.toLowerCase()) {
     case "document":
       return OfficeApp.Word;
+    case "mail":
+      return OfficeApp.Outlook;
     case "mailbox":
       return OfficeApp.Outlook;
     case "notebook":
@@ -70,7 +72,7 @@ export function getOfficeAppName(app: OfficeApp): string {
  * @param apps Office apps
  */
 export function getOfficeAppNames(apps: OfficeApp[]): string[] {
-  return apps.map(app => getOfficeAppName(app));
+  return apps.map((app) => getOfficeAppName(app));
 }
 
 /**
@@ -88,7 +90,7 @@ export function getOfficeAppsForManifestHosts(hosts?: string[]): OfficeApp[] {
   const apps: OfficeApp[] = [];
 
   if (hosts) {
-    hosts.forEach(host => {
+    hosts.forEach((host) => {
       const app = getOfficeAppForManifestHost(host);
 
       if (app) {
@@ -124,7 +126,7 @@ export function parseOfficeApps(input: string): OfficeApp[] {
   if (input.trim().toLowerCase() === "all") {
     return getOfficeApps();
   } else {
-    return input.split(",").map<OfficeApp>(appString => parseOfficeApp(appString));
+    return input.split(",").map<OfficeApp>((appString) => parseOfficeApp(appString));
   }
 }
 
