@@ -6,14 +6,14 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-import * as commander from 'commander';
-import { run } from './debugger';
+import * as commander from "commander";
+import { run } from "./debugger";
+
+/* global process */
 
 commander
-.option('-h, --host <host>', 'The hostname where the packager is running.')
-.option('-p, --port <port>', 'The port where the packager is running.')  
-.parse(process.argv);
+  .option("-h, --host <host>", "The hostname where the packager is running.")
+  .option("-p, --port <port>", "The port where the packager is running.")
+  .parse(process.argv);
 
 run(commander.host, commander.port);
-
-
