@@ -3,30 +3,28 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import * as commander from 'commander';
-import { logErrorMessage } from 'office-addin-cli';
-import * as commands from './commands';
+import * as commander from "commander";
+import { logErrorMessage } from "office-addin-usage-data";
+import * as commands from "./commands";
 
-commander.name('office-addin-sso');
-commander.version(process.env.npm_package_version || '(version not available)');
+/* global process */
 
-commander
-    .command('configure <manifest-path>')
-    .action(commands.configureSSO);
+commander.name("office-addin-sso");
+commander.version(process.env.npm_package_version || "(version not available)");
 
-commander
-    .command('start <manifest-path>')
-    .action(commands.startSSOService);
+commander.command("configure <manifest-path>").action(commands.configureSSO);
+
+commander.command("start <manifest-path>").action(commands.startSSOService);
 
 // if the command is not known, display an error
-commander.on('command:*', () => {
-    logErrorMessage(`The command syntax is not valid.\n`);
-    process.exitCode = 1;
-    commander.help();
+commander.on("command:*", () => {
+  logErrorMessage(`The command syntax is not valid.\n`);
+  process.exitCode = 1;
+  commander.help();
 });
 
 if (process.argv.length > 2) {
-    commander.parse(process.argv);
+  commander.parse(process.argv);
 } else {
-    commander.help();
+  commander.help();
 }
