@@ -1,20 +1,32 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import * as path from 'path';
+import * as path from "path";
 import * as usageData from "office-addin-usage-data";
+
+/* global process, __dirname */
 
 // File path defaults
 export const azCliInstallCommandPath: string = path.resolve(`${__dirname}/scripts/azCliInstallCmd.ps1`);
-export const azRestAddTenantOutlookReplyUrlsCommandPath = path.resolve(`${__dirname}/scripts/azRestAddTenantOutlookReplyUrls.txt`);
+export const azRestAddTenantOutlookReplyUrlsCommandPath = path.resolve(
+  `${__dirname}/scripts/azRestAddTenantOutlookReplyUrls.txt`
+);
 export const azRestAddTenantReplyUrlsCommandPath = path.resolve(`${__dirname}/scripts/azRestAddTenantReplyUrls.txt`);
 export const azRestAddSecretCommandPath = path.resolve(`${__dirname}/scripts/azAddSecretCmd.txt`);
 export const azRestAppCreateCommandPath: string = path.resolve(`${__dirname}/scripts/azRestAppCreateCmd.txt`);
-export const azRestGetOrganizationDetailsCommandPath: string = path.resolve(`${__dirname}/scripts/azGetOrganizationDetails.txt`);
-export const azRestGetTenantAdminMembershipCommandPath: string = path.resolve(`${__dirname}/scripts/azRestGetTenantAdminMembership.txt`);
+export const azRestGetOrganizationDetailsCommandPath: string = path.resolve(
+  `${__dirname}/scripts/azGetOrganizationDetails.txt`
+);
+export const azRestGetTenantAdminMembershipCommandPath: string = path.resolve(
+  `${__dirname}/scripts/azRestGetTenantAdminMembership.txt`
+);
 export const azRestGetTenantRolesPath: string = path.resolve(`${__dirname}/scripts/azRestGetTenantRoles.txt`);
-export const azRestSetIdentifierUriCommmandPath: string = path.resolve(`${__dirname}/scripts/azRestSetIdentifierUri.txt`);
-export const azRestSetSigninAudienceCommandPath: string = path.resolve(`${__dirname}/scripts/azSetSignInAudienceCmd.txt`);
+export const azRestSetIdentifierUriCommmandPath: string = path.resolve(
+  `${__dirname}/scripts/azRestSetIdentifierUri.txt`
+);
+export const azRestSetSigninAudienceCommandPath: string = path.resolve(
+  `${__dirname}/scripts/azSetSignInAudienceCmd.txt`
+);
 export const envDataFilePath = path.resolve(`${process.cwd()}/.ENV`);
 export const fallbackAuthDialogTypescriptFilePath = path.resolve(`${process.cwd()}/src/helpers/fallbackAuthDialog.ts`);
 export const fallbackAuthDialogJavascriptFilePath = path.resolve(`${process.cwd()}/src/helpers/fallbackAuthDialog.js`);
@@ -28,7 +40,7 @@ export const testManifestFilePath = path.resolve(`${process.cwd()}/test/test-man
 
 // Usage data defaults
 export const usageDataObject: usageData.OfficeAddinUsageData = new usageData.OfficeAddinUsageData({
-    projectName: "office-addin-sso",
-    instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
-    raisePrompt: false
+  projectName: "office-addin-sso",
+  instrumentationKey: usageData.instrumentationKeyForOfficeAddinCLITools,
+  raisePrompt: false,
 });
