@@ -1,5 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-export * from './error-handler';
-export * from './message-helper';
-export * from './msgraph-helper';
+export * from "./error-handler";
+export * from "./message-helper";
+export * from "./msgraph-helper";
