@@ -13,13 +13,10 @@ const testValue: string = "Office-Addin-Test-Infrastructure";
 const testValues: any = [];
 
 describe("Start test server, validate pingTestServer and sendTestResults methods and stop test server", function () {
-    before(function() {
-        if (process.platform == "linux") {
-            this.skip();
-        }
-    });
     describe("Setup test server", function () {
         it("Test server should have started", async function () {
+            // give the server some time to start
+            this.timeout(10000);
             const startTestServer = await promiseStartTestServer;
             assert.equal(startTestServer, true);
         });
