@@ -2,27 +2,29 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
-import { logErrorMessage } from "office-addin-cli";
+import { logErrorMessage } from "office-addin-usage-data";
 import * as commands from "./commands";
 import { defaultPort } from "./testServer";
+
+/* global process */
 
 commander.name("office-addin-test-server");
 commander.version(process.env.npm_package_version || "(version not available)");
 
 commander
-    .command("start")
-    .option(`-p --port [port number]", "Port number must be between 0 - 65535. Default: ${defaultPort}`)
-    .action(commands.start);
+  .command("start")
+  .option(`-p --port [port number]", "Port number must be between 0 - 65535. Default: ${defaultPort}`)
+  .action(commands.start);
 
 // if the command is not known, display an error
-commander.on("command:*", function() {
-    logErrorMessage(`The command syntax is not valid.\n`);
-    process.exitCode = 1;
-    commander.help();
+commander.on("command:*", function () {
+  logErrorMessage(`The command syntax is not valid.\n`);
+  process.exitCode = 1;
+  commander.help();
 });
 
 if (process.argv.length > 2) {
-    commander.parse(process.argv);
+  commander.parse(process.argv);
 } else {
-    commander.help();
+  commander.help();
 }
